Improve sorting tasks, Add feature: Cut tasks that are done, Minor Bug Fix
