+++ To-do-list/temp_list/orig_wallpaper.txt
@@ -1,1 +1,1 @@
-/usr/share/backgrounds/Green_Island_by_Radu_Galan.jpg
+/usr/share/backgrounds/Lighthouse_at_sunrise_by_Frenchie_Smalls.jpg
